conda_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
aiofiles=23.2.1=pypi_0
altair=5.2.0=pypi_0
annotated-types=0.6.0=py311haa95532_0
aom=3.6.1=h63175ca_0
blas=1.0=mkl
bottleneck=1.3.7=py311hd7041d2_0
brotli=1.0.9=h2bbff1b_7
brotli-bin=1.0.9=h2bbff1b_7
brotli-python=1.0.9=py311hd77b12b_7
bzip2=1.0.8=h2bbff1b_5
ca-certificates=2024.3.11=haa95532_0
cairo=1.16.0=haedb8bc_5
catalogue=2.0.10=py311haa95532_0
certifi=2024.2.2=py311haa95532_0
cffi=1.16.0=py311h2bbff1b_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.7=py311haa95532_0
cloudpathlib=0.16.0=py311haa95532_1
colorama=0.4.6=py311haa95532_0
confection=0.1.4=py311h746a85d_0
contourpy=1.2.0=py311h59b6b97_0
cuda-cccl=12.4.99=0
cuda-cudart=12.1.105=0
cuda-cudart-dev=12.1.105=0
cuda-cupti=12.1.105=0
cuda-libraries=12.1.0=0
cuda-libraries-dev=12.1.0=0
cuda-nvrtc=12.1.105=0
cuda-nvrtc-dev=12.1.105=0
cuda-nvtx=12.1.105=0
cuda-opencl=12.4.99=0
cuda-opencl-dev=12.4.99=0
cuda-profiler-api=12.4.99=0
cuda-runtime=12.1.0=0
cudatoolkit=11.8.0=hd77b12b_0
cudnn=8.9.2.26=cuda11_0
cycler=0.11.0=pyhd3eb1b0_0
cymem=2.0.6=py311hd77b12b_0
cython-blis=0.7.9=py311hd7041d2_0
dav1d=1.2.1=h2bbff1b_0
expat=2.5.0=hd77b12b_0
fastai=2.7.14=py_0
fastcore=1.5.29=py311haa95532_1
fastdownload=0.0.7=py311haa95532_0
fastprogress=1.0.0=pyhb85f177_0
ffmpeg=5.1.2=gpl_h8bb4bc8_112
ffmpy=0.3.2=pypi_0
filelock=3.13.1=py311haa95532_0
font-ttf-dejavu-sans-mono=2.37=hd3eb1b0_0
font-ttf-inconsolata=2.001=hcb22688_0
font-ttf-source-code-pro=2.030=hd3eb1b0_0
font-ttf-ubuntu=0.83=h8b1ccd4_0
fontconfig=2.14.2=hbde0cde_0
fonts-anaconda=1=h8fa9717_0
fonts-conda-ecosystem=1=hd3eb1b0_0
fonttools=4.25.0=pyhd3eb1b0_0
freeglut=3.2.2=h63175ca_2
freetype=2.12.1=ha860e81_0
fsspec=2023.10.0=py311haa95532_0
glib=2.78.4=hd77b12b_0
glib-tools=2.78.4=hd77b12b_0
gmpy2=2.1.2=py311h7f96b67_0
gradio=4.24.0=pypi_0
gradio-client=0.14.0=pypi_0
graphite2=1.3.14=hd77b12b_1
gst-plugins-base=1.22.3=h23ce68f_0
gstreamer=1.22.3=h2bbff1b_0
h11=0.14.0=pypi_0
harfbuzz=6.0.0=he256f1b_0
httpcore=1.0.5=pypi_0
httpx=0.27.0=pypi_0
huggingface-hub=0.22.1=pypi_0
icc_rt=2022.1.0=h6049295_2
icu=70.1=h0e60522_0
idna=3.4=py311haa95532_0
importlib-resources=6.4.0=pypi_0
intel-openmp=2023.1.0=h59b6b97_46320
jasper=2.0.33=hc2e4405_1
jinja2=3.1.3=py311haa95532_0
joblib=1.2.0=py311haa95532_0
jpeg=9e=h2bbff1b_1
jsonschema=4.21.1=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
kiwisolver=1.4.4=py311hd77b12b_0
krb5=1.20.1=h5b6d351_0
langcodes=3.3.0=pyhd3eb1b0_0
lerc=3.0=hd77b12b_0
libblas=3.9.0=1_h8933c1f_netlib
libbrotlicommon=1.0.9=h2bbff1b_7
libbrotlidec=1.0.9=h2bbff1b_7
libbrotlienc=1.0.9=h2bbff1b_7
libcblas=3.9.0=5_hd5c7e75_netlib
libclang=15.0.7=default_h3a3e6c3_5
libclang13=15.0.7=default_hf64faad_5
libcublas=12.1.0.26=0
libcublas-dev=12.1.0.26=0
libcufft=11.0.2.4=0
libcufft-dev=11.0.2.4=0
libcurand=10.3.5.119=0
libcurand-dev=10.3.5.119=0
libcusolver=11.4.4.55=0
libcusolver-dev=11.4.4.55=0
libcusparse=12.0.2.55=0
libcusparse-dev=12.0.2.55=0
libdeflate=1.17=h2bbff1b_1
libffi=3.4.4=hd77b12b_0
libglib=2.78.4=ha17d25a_0
libiconv=1.17=hcfcfb64_2
libjpeg-turbo=2.0.0=h196d8e1_0
liblapack=3.9.0=5_hd5c7e75_netlib
liblapacke=3.9.0=5_hd5c7e75_netlib
libnpp=12.0.2.50=0
libnpp-dev=12.0.2.50=0
libnvjitlink=12.1.105=0
libnvjitlink-dev=12.1.105=0
libnvjpeg=12.1.1.14=0
libnvjpeg-dev=12.1.1.14=0
libogg=1.3.5=h2bbff1b_1
libopencv=4.7.0=py311h0442f2a_1
libopus=1.3.1=h8ffe710_1
libpng=1.6.39=h8cc25b3_0
libpq=12.17=h906ac69_0
libprotobuf=3.21.12=h12be248_2
libsqlite=3.45.2=hcfcfb64_0
libtiff=4.5.1=hd77b12b_0
libuv=1.44.2=h2bbff1b_0
libvorbis=1.3.7=he774522_0
libwebp-base=1.3.2=h2bbff1b_0
libxml2=2.12.6=hc3477c8_1
libzlib=1.2.13=hcfcfb64_5
lz4-c=1.9.4=h2bbff1b_0
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markdown-it-py=2.2.0=py311haa95532_1
markupsafe=2.1.3=py311h2bbff1b_0
matplotlib=3.8.0=py311haa95532_0
matplotlib-base=3.8.0=py311hf62ec03_0
mdurl=0.1.0=py311haa95532_0
mkl=2023.1.0=h6b88ed4_46358
mkl-service=2.4.0=py311h2bbff1b_1
mkl_fft=1.3.8=py311h2bbff1b_0
mkl_random=1.2.4=py311h59b6b97_0
mpc=1.1.0=h7edee0f_1
mpfr=4.0.2=h62dcd97_1
mpir=3.0.0=hec2e145_1
mpmath=1.3.0=py311haa95532_0
msys2-conda-epoch=20160418=1
munkres=1.1.4=py_0
murmurhash=1.0.7=py311hd77b12b_0
networkx=3.1=py311haa95532_0
ninja=1.10.2=haa95532_5
ninja-base=1.10.2=h6d14046_5
numexpr=2.8.7=py311h1fcbade_0
numpy=1.24.3=py311hdab7c0b_1
numpy-base=1.24.3=py311hd01c5d8_1
opencv=4.7.0=py311h1ea47a8_1
opencv-contrib-python=4.9.0.80=pypi_0
openh264=2.3.1=h63175ca_2
openjpeg=2.4.0=h4fc8c34_0
openssl=3.2.1=hcfcfb64_1
orjson=3.10.0=pypi_0
packaging=23.1=py311haa95532_0
pandas=2.2.1=py311hea22821_0
pcre2=10.42=h0ff8eda_0
pillow=10.2.0=py311h2bbff1b_0
pip=23.3.1=py311haa95532_0
pixman=0.40.0=h2bbff1b_1
ply=3.11=py311haa95532_0
preshed=3.0.6=py311h6c2663c_0
py-opencv=4.7.0=py311h3810d55_1
pycparser=2.21=pyhd3eb1b0_0
pydantic=2.5.3=py311haa95532_0
pydantic-core=2.14.6=py311h062c2fa_0
pydub=0.25.1=pypi_0
pygments=2.15.1=py311haa95532_1
pyparsing=3.0.9=py311haa95532_0
pyqt=5.15.10=py311hd77b12b_0
pyqt5-sip=12.13.0=py311h2bbff1b_0
pysocks=1.7.1=py311haa95532_0
python=3.11.8=he1021f5_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-multipart=0.0.9=pypi_0
python-tzdata=2023.3=pyhd3eb1b0_0
python_abi=3.11=2_cp311
pytorch=2.2.1=py3.11_cuda12.1_cudnn8_0
pytorch-cuda=12.1=hde6ce7c_5
pytorch-mutex=1.0=cuda
pytz=2023.3.post1=py311haa95532_0
pyyaml=6.0.1=py311h2bbff1b_0
qt-main=5.15.8=h720456b_6
referencing=0.34.0=pypi_0
requests=2.31.0=py311haa95532_1
rich=13.3.5=py311haa95532_0
rpds-py=0.18.0=pypi_0
ruff=0.3.4=pypi_0
scikit-learn=1.3.0=py311hf62ec03_1
scipy=1.11.4=py311hc1ccb85_0
seaborn=0.12.2=py311haa95532_0
semantic-version=2.10.0=pypi_0
setuptools=68.2.2=py311haa95532_0
shellingham=1.5.0=py311haa95532_0
sip=6.7.12=py311hd77b12b_0
six=1.16.0=pyhd3eb1b0_1
smart_open=5.2.1=py311haa95532_0
spacy=3.7.2=py311ha84f9a1_0
spacy-legacy=3.0.12=py311haa95532_0
spacy-loggers=1.0.4=py311haa95532_0
sqlite=3.41.2=h2bbff1b_0
srsly=2.4.8=py311hd77b12b_1
svt-av1=1.4.1=h63175ca_0
sympy=1.12=py311haa95532_0
tbb=2021.8.0=h59b6b97_0
thinc=8.2.2=py311h30df693_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.12=h2bbff1b_0
tomlkit=0.12.0=pypi_0
toolz=0.12.1=pypi_0
torchaudio=2.2.1=pypi_0
torchvision=0.17.1=pypi_0
tornado=6.3.3=py311h2bbff1b_0
tqdm=4.65.0=py311h746a85d_0
typer=0.9.0=py311haa95532_0
typing-extensions=4.9.0=py311haa95532_1
typing_extensions=4.9.0=py311haa95532_1
tzdata=2024a=h04d1e81_0
ucrt=10.0.20348.0=haa95532_0
urllib3=2.1.0=py311haa95532_1
uvicorn=0.29.0=pypi_0
vc=14.2=h21ff451_1
vc14_runtime=14.38.33130=h82b7239_18
vs2015_runtime=14.38.33130=hcb4865c_18
wasabi=0.9.1=py311haa95532_0
weasel=0.3.4=py311haa95532_0
websockets=11.0.3=pypi_0
wheel=0.41.2=py311haa95532_0
win_inet_pton=1.1.0=py311haa95532_0
x264=1!164.3095=h8ffe710_2
x265=3.5=h2d74725_3
xz=5.4.6=h8cc25b3_0
yaml=0.2.5=he774522_0
zlib=1.2.13=hcfcfb64_5
zstd=1.5.5=hd43e919_0
